--- src/main/java/com/goalapa/cacamuca/foodDomain/command/domain/repository/BlackListRepository.java
@@ -1,10 +1,0 @@
-package com.goalapa.cacamuca.foodDomain.command.domain.repository;
-
-import com.goalapa.cacamuca.blackListDomain.command.domain.aggregate.entity.Blacklist;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BlackListRepository extends JpaRepository<Blacklist, Integer> {
-
-}
