+++ src/main/java/com/goalapa/cacamuca/foodDomain/command/ininfrastructure/repository/Repository.java
@@ -1,0 +1,4 @@
+package com.goalapa.cacamuca.foodDomain.command.ininfrastructure.repository;
+
+public class Repository {
+}
