+++ src/main/java/com/goalapa/cacamuca/memberDomain/command/application/controller/CommandMemberController.java
@@ -1,0 +1,12 @@
+package com.goalapa.cacamuca.memberDomain.command.application.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/member1")
+public class CommandMemberController {
+
+}
