--- src/main/java/com/goalapa/cacamuca/memberDomain/command/application/controller/Controller.java
@@ -1,4 +1,0 @@
-package com.goalapa.cacamuca.memberDomain.command.application.controller;
-
-public class Controller {
-}
