--- src/main/java/com/goalapa/cacamuca/memberDomain/command/model/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.goalapa.cacamuca.memberDomain.command.model.repository;
-
-import com.goalapa.cacamuca.memberDomain.command.domain.aggregate.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Integer> {
-}
