--- src/main/java/com/goalapa/cacamuca/memberDomain/query/application/controller/Controller.java
@@ -1,4 +1,0 @@
-package com.goalapa.cacamuca.memberDomain.query.application.controller;
-
-public class Controller {
-}
