+++ src/main/java/com/goalapa/cacamuca/memberDomain/query/application/controller/MemberController.java
@@ -1,7 +1,7 @@
-package com.goalapa.cacamuca.memberDomain.command.model.controller;
+package com.goalapa.cacamuca.memberDomain.query.application.controller;
 
 import com.goalapa.cacamuca.memberDomain.command.domain.aggregate.entity.Member;
-import com.goalapa.cacamuca.memberDomain.command.model.service.MemberService;
+import com.goalapa.cacamuca.memberDomain.query.application.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,17 +12,16 @@ import java.util.List;
 import java.util.Map;
 
 @Controller
-@RequestMapping("/member")
 @RequiredArgsConstructor
+@RequestMapping("/member")
 public class MemberController {
 
     private final MemberService memberService;
 
-    @GetMapping("/memberList")
-    @ResponseBody
-    public List<Member> getMemberList() {
+    @GetMapping("/main")
+    public String main() {
 
-        return memberService.getMemberList();
+        return "/main";
     }
 
     @GetMapping("/member")
