+++ src/main/resources/static/index.html
@@ -5,6 +5,6 @@
     <title>Title</title>
 </head>
 <body>
-    <button onclick="location.href='/main'">메인</button>
+    <button onclick="location.href='/member/main'">메인</button>
 </body>
 </html>
